+++ src/components/themed/SideMenu.tsx
@@ -33,6 +33,7 @@ import { navigateToGiftCards } from '../../actions/GiftCardActions'
 import { useNotifCount } from '../../actions/LocalSettingsActions'
 import { getRootNavigation, logoutRequest } from '../../actions/LoginActions'
 import { executePluginAction } from '../../actions/PluginActions'
+import { showScanModal } from '../../actions/ScanActions'
 import { Fontello } from '../../assets/vector'
 import { SCROLL_INDICATOR_INSET_FIX } from '../../constants/constantSettings'
 import { ENV } from '../../env'
@@ -52,7 +53,6 @@ import { styled } from '../hoc/styled'
 import { IconBadge } from '../icons/IconBadge'
 import { ChevronDownIcon, CloseIcon } from '../icons/ThemedIcons'
 import { ButtonsModal } from '../modals/ButtonsModal'
-import { ScanModal } from '../modals/ScanModal'
 import { Airship, showError } from '../services/AirshipInstance'
 import { Services } from '../services/Services'
 import { cacheStyles, type Theme, useTheme } from '../services/ThemeContext'
@@ -164,16 +164,15 @@ export function SideMenuComponent(props: Props): React.ReactElement {
 
   const handleScanQr = (): void => {
     navigation.dispatch(DrawerActions.closeDrawer())
-    Airship.show<string | undefined>(bridge => (
-      <ScanModal
-        bridge={bridge}
-        scanModalTitle={lstrings.scan_qr_label}
-        textModalAutoFocus={false}
-        textModalTitle={lstrings.enter_any_title}
-        textModalBody={lstrings.enter_any_body}
-        textModalHint={lstrings.enter_any_input_hint}
-      />
-    ))
+    dispatch(
+      showScanModal({
+        scanModalTitle: lstrings.scan_qr_label,
+        textModalAutoFocus: false,
+        textModalTitle: lstrings.enter_any_title,
+        textModalBody: lstrings.enter_any_body,
+        textModalHint: lstrings.enter_any_input_hint
+      })
+    )
       .then(async (result: string | undefined) => {
         if (result != null && result !== '') {
           const deepLink = parseDeepLink(result)